Maintain backwards compatibility
We need a way to know whether the loaded file is one supporting the
custom backend or not. This is done by placing a root key `_`. When this
is present, the file is parsed using the custom logic, otherwise it
falls back to the logic of the simple backend.

Without this addition, normal files would fail to load.